pyproject.toml
[tool.poetry]
name = "cfbm"
version = "2022.4"
description = "CHIME/FRB Beam Model"
authors = ["Paul Scholz <paul.scholz@dunlap.utoronto.ca>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
pyephem = "^3.7.7"
numpy = "^1.18.5"
scipy = "^1.4.1"
h5py = "^2.10.0"
requests = "^2.24.0"
chime-frb-constants = "^2020.07"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
pytz = "^2020.1"
pytest-cov = "^2.10.0"
coveralls = {version = "^2.1.1", extras = ["toml"]}
toml = "^0.10.1"

[tool.poetry.scripts]
get-data = "cfbm.bm_data.get_data:main"

[tool.coverage.run]
omit = ["**/get_data.py"]
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

